+++ features/Layout/components/layout/Layout.tsx
@@ -2,84 +2,30 @@ import { FunctionComponent, ReactNode } from "react";
 
 import Head from "next/head";
 
-import { uppercaseFirst } from "core";
+import { Flex } from "@chakra-ui/react";
 
-import { Footer, Navbar } from "..";
+import LayoutMain, { LayoutMainType } from "./layout-main/LayoutMain";
 
-import {
-  FAVICON_URL,
-  SITE_DESCRIPTION,
-  SITE_IMAGE,
-  SITE_NAME,
-} from "core/constants";
-
-import styles from "./Layout.module.scss";
+interface LayoutComposition {
+  Head: ({ children }: { children: ReactNode }) => JSX.Element;
+  Main: LayoutMainType;
+}
 
 interface LayoutProps {
   children: ReactNode;
-  title?: string;
-  siteImage?: string;
-  metaDescription?: string;
 }
 
-const Layout: FunctionComponent<LayoutProps> = ({
-  children,
-  title,
-  siteImage,
-  metaDescription,
-}) => {
-  const siteUrl = process.env.NEXT_PUBLIC_SITE_URL || "";
+type LayoutType = FunctionComponent<LayoutProps> & LayoutComposition;
 
-  const twitterImageContentUrl = siteImage ? siteImage : SITE_IMAGE;
-  const twitterImageContent = `${siteUrl}${twitterImageContentUrl}`;
-
-  const pageTitle = title
-    ? `${uppercaseFirst(title)} - ${SITE_NAME}`
-    : SITE_NAME;
-  const pageDescription = metaDescription
-    ? metaDescription
-    : `${SITE_NAME} - ${SITE_DESCRIPTION}`;
-
-  const propertiesTitle = title ? uppercaseFirst(title) : SITE_NAME;
-  const propertiesDescription = metaDescription
-    ? metaDescription
-    : SITE_DESCRIPTION;
-
+const Layout: LayoutType = ({ children }) => {
   return (
-    <div className={styles.container}>
-      <Head>
-        <title>{pageTitle}</title>
-        <meta name="description" content={pageDescription} />
-
-        {/* Google Properties */}
-        <meta itemProp="name" content={propertiesTitle} />
-        <meta itemProp="description" content={propertiesDescription} />
-        <meta itemProp="image" content={siteImage ? siteImage : SITE_IMAGE} />
-
-        {/* Open Graph Properties */}
-        <meta property="og:url" content="https://www.devoreur2code.com" />
-        <meta property="og:type" content="website" />
-        <meta property="og:title" content={propertiesTitle} />
-        <meta property="og:description" content={propertiesDescription} />
-        <meta
-          property="og:image"
-          content={siteImage ? siteImage : SITE_IMAGE}
-        />
-
-        {/* Twitter Properties */}
-        <meta name="twitter:card" content="summary" />
-        <meta name="twitter:site" content="@devoreur2code" />
-        <meta name="twitter:title" content={propertiesTitle} />
-        <meta name="twitter:description" content={propertiesDescription} />
-        <meta name="twitter:image" content={twitterImageContent} />
-
-        <link rel="icon" href={FAVICON_URL} />
-      </Head>
-      <Navbar />
-      <main className={styles.main}>{children}</main>
-      <Footer />
-    </div>
+    <Flex display="flex" flexDirection="column" minHeight="100vh" padding={0}>
+      {children}
+    </Flex>
   );
 };
+
+Layout.Head = Head;
+Layout.Main = LayoutMain;
 
 export default Layout;