+++ features/Layout/hooks/index.ts
@@ -1,0 +1,3 @@
+import useMetaProperties from "./use-meta-properties/useMetaProperties";
+
+export { useMetaProperties };